Use separate files for the workshops and projects API endpoints
